fix agent tests
