+++ examples/multi_agent_human_chat/Makefile
@@ -1,6 +1,6 @@
 .PHONY: default setup start test test-billing-agent test-claims-agent test-escalation-agent test-frontend-agent test-policies-agent test-policies-retrieval-performance test-triage-classifier-v0 test-triage-classifier-v1 test-triage-classifier-v2 test-triage-classifier-v3 test-triage-classifier-v4 compile-triage-classifier-v2 compile-triage-classifier-v4 compile-billing-optimizer compile-claims-optimizer compile-policies-optimizer compile-escalation-optimizer compile-all test-triage-agent test-audit test-libraries stop kill-agents clean full-reset docker-up docker-down start-frontend start-billing start-escalation start-policies start-claims start-triage start-audit start-policies-document-ingestion start-all restart setup-and-run lint lint-fix build-policy-rag-index drop-vespa-index deploy-vespa-schema
 
-PYTHON := python3.11
+PYTHON ?= python3
 VENV_DIR := .venv
 REQ_FILE := requirements.txt
 DEV_REQ_FILE := dev-requirements.txt
@@ -208,7 +208,7 @@ start-escalation:
 
 start-policies:
 	@echo "Starting Policies Agent..."
-	@source $(VENV_DIR)/bin/activate && $(VENV_DIR)/bin/python -m agents.policies.agent.main
+	@source $(VENV_DIR)/bin/activate && $(VENV_DIR)/bin/python -m scripts.start_policies
 
 start-policies-document-ingestion:
 	@echo "Starting Policies Document Ingestion Service..."
