+++ examples/multi_agent_human_chat/agents/audit/agent.py
@@ -1,5 +1,5 @@
+import asyncio
 from typing import Dict, Optional, Union
-from uuid import uuid4
 
 from eggai import Agent, Channel
 from eggai.transport import eggai_set_default_transport
@@ -8,7 +8,8 @@ from faststream.kafka import KafkaMessage
 from libraries.channels import channels
 from libraries.kafka_transport import create_kafka_transport
 from libraries.logger import get_console_logger
-from libraries.tracing import TracedMessage, create_tracer, traced_handler
+from libraries.trace_utils import get_message_id, get_message_metadata
+from libraries.tracing import TracedMessage, get_tracer, traced_handler
 from libraries.tracing.init_metrics import init_token_metrics
 from libraries.tracing.otel import safe_set_attribute
 
@@ -22,11 +23,7 @@ eggai_set_default_transport(
     )
 )
 logger = get_console_logger("audit_agent")
-
-agents_channel = Channel("agents")
-human_channel = Channel("human")
-audit_logs_channel = Channel("audit_logs")
-tracer = create_tracer("audit_agent")
+tracer = get_tracer("audit_agent")
 
 init_token_metrics(
     port=settings.prometheus_metrics_port, application_name=settings.app_name
@@ -48,71 +45,10 @@ audit_config = AuditConfig(
 )
 
 audit_agent = Agent("AuditAgent")
-logger = get_console_logger("audit_agent")
 
 agents_channel = Channel(channels.agents)
 human_channel = Channel(channels.human)
 audit_logs_channel = Channel(channels.audit_logs)
-
-
-def get_message_metadata(
-    message: Optional[Union[TracedMessage, Dict]],
-) -> tuple[str, str]:
-    if message is None:
-        return "unknown", "unknown"
-
-    if hasattr(message, "type") and hasattr(message, "source"):
-        return message.type, message.source
-
-    try:
-        return message.get("type", "unknown"), message.get("source", "unknown")
-    except (AttributeError, TypeError):
-        logger.warning("Message has no type or source attributes and is not dict-like")
-        return "unknown", "unknown"
-
-
-def get_message_content(message: Optional[Union[TracedMessage, Dict]]) -> Optional[str]:
-    """Extract message content from various message formats."""
-    if message is None:
-        return None
-
-    try:
-        if hasattr(message, "data"):
-            data = message.data
-            if isinstance(data, dict):
-                if "message" in data:
-                    return data["message"]
-                if (
-                    "chat_messages" in data
-                    and isinstance(data["chat_messages"], list)
-                    and data["chat_messages"]
-                ):
-                    # Get the last user message from chat history
-                    last_msg = data["chat_messages"][-1]
-                    if isinstance(last_msg, dict) and "content" in last_msg:
-                        return last_msg["content"]
-        return None
-    except (AttributeError, TypeError, IndexError):
-        logger.warning("Could not extract message content", exc_info=True)
-        return None
-
-
-def get_message_id(message: Optional[Union[TracedMessage, Dict]]) -> str:
-    if message is None:
-        return f"null_message_{uuid4()}"
-    return str(getattr(message, "id", uuid4()))
-
-
-def propagate_trace_context(
-    source_message: Optional[Union[TracedMessage, Dict]], target_message: TracedMessage
-) -> None:
-    if source_message is None:
-        return
-
-    if hasattr(source_message, "traceparent") and source_message.traceparent:
-        target_message.traceparent = source_message.traceparent
-    if hasattr(source_message, "tracestate") and source_message.tracestate:
-        target_message.tracestate = source_message.tracestate
 
 
 @audit_agent.subscribe(channel=agents_channel)
@@ -151,6 +87,14 @@ async def audit_message(
 
         return message
 
-    except Exception as e:
-        logger.error(f"Error processing audit message: {e}", exc_info=True)
+    except (KeyError, ValueError) as exc:
+        logger.warning("Invalid audit message: %s", exc, exc_info=True)
+        return message
+    except (ConnectionError, asyncio.TimeoutError) as exc:
+        logger.error("Network error processing audit message: %s", exc, exc_info=True)
+        return message
+    except Exception as exc:
+        logger.error(
+            "Unexpected error processing audit message: %s", exc, exc_info=True
+        )
         return message
