+++ examples/multi_agent_human_chat/agents/billing/main.py
@@ -46,5 +46,7 @@ if __name__ == "__main__":
         asyncio.run(main())
     except KeyboardInterrupt:
         logger.info("Shutting down billing agent")
+    except OSError as e:
+        logger.error(f"Server error: {e}", exc_info=True)
     except Exception as e:
         logger.error(f"Error in main: {e}", exc_info=True)
