+++ examples/multi_agent_human_chat/agents/claims/agent.py
@@ -1,19 +1,18 @@
 import asyncio
 from typing import List
 
-from dspy import Prediction
-from dspy.streaming import StreamResponse
 from eggai import Agent, Channel
+from opentelemetry.propagate import TraceContextTextMapPropagator
 
 from agents.claims.config import settings
 from agents.claims.dspy_modules.claims import claims_optimized_dspy
 from agents.claims.types import ChatMessage, ModelConfig
-from libraries.channels import channels, clear_channels
+from libraries.channels import channels
 from libraries.logger import get_console_logger
+from libraries.streaming_agent import format_conversation, process_request_stream
 from libraries.tracing import (
     TracedMessage,
-    create_tracer,
-    format_span_as_traceparent,
+    get_tracer,
     traced_handler,
 )
 from libraries.tracing.init_metrics import init_token_metrics
@@ -24,7 +23,7 @@ logger = get_console_logger("claims_agent.handler")
 agents_channel = Channel(channels.agents)
 human_channel = Channel(channels.human)
 human_stream_channel = Channel(channels.human_stream)
-tracer = create_tracer("claims_agent")
+tracer = get_tracer("claims_agent")
 
 init_token_metrics(
     port=settings.prometheus_metrics_port, application_name=settings.app_name
@@ -32,41 +31,37 @@ init_token_metrics(
 
 
 def get_conversation_string(chat_messages: List[ChatMessage]) -> str:
-    """Format chat messages into a conversation string."""
-    with tracer.start_as_current_span("get_conversation_string") as span:
-        safe_set_attribute(
-            span, "chat_messages_count", len(chat_messages) if chat_messages else 0
-        )
-
-        if not chat_messages:
-            safe_set_attribute(span, "empty_messages", True)
-            return ""
-
-        conversation_parts = []
-        for chat in chat_messages:
-            if "content" not in chat:
-                safe_set_attribute(span, "invalid_message", True)
-                continue
-
-            role = chat.get("role", "User")
-            conversation_parts.append(f"{role}: {chat['content']}")
-
-        conversation = "\n".join(conversation_parts) + "\n"
-        safe_set_attribute(span, "conversation_length", len(conversation))
-        return conversation
+    """Legacy wrapper for tests that formats chat history."""
+    return format_conversation(chat_messages, tracer=tracer, logger=logger)
 
 
 async def process_claims_request(
     conversation_string: str,
     connection_id: str,
     message_id: str,
-    timeout_seconds: float = None,
+    timeout_seconds: float | None = None,
 ) -> None:
-    """Generate a response to a claims request with streaming output."""
+    """Stream a claims response back to the user.
+
+    Args:
+        conversation_string: Formatted conversation history.
+        connection_id: Identifier of the user's connection.
+        message_id: Unique ID for this claims request.
+        timeout_seconds: Optional model timeout override.
+
+    Returns:
+        None
+    """
     # Create model config with timeout value
-    config = ModelConfig(name="claims_react", timeout_seconds=timeout_seconds or 30.0)
+    config = ModelConfig(
+        name="claims_react",
+        timeout_seconds=timeout_seconds or settings.request_timeout_seconds,
+    )
     with tracer.start_as_current_span("process_claims_request") as span:
-        child_traceparent, child_tracestate = format_span_as_traceparent(span)
+        carrier = {}
+        TraceContextTextMapPropagator().inject(carrier)
+        child_traceparent = carrier.get("traceparent")
+        child_tracestate = carrier.get("tracestate", "")
         safe_set_attribute(span, "connection_id", connection_id)
         safe_set_attribute(span, "message_id", message_id)
         safe_set_attribute(span, "conversation_length", len(conversation_string))
@@ -77,86 +72,20 @@ async def process_claims_request(
             span.set_status(1, "Invalid input")
             raise ValueError("Conversation history is too short to process")
 
-        # Start the stream
-        await human_stream_channel.publish(
-            TracedMessage(
-                type="agent_message_stream_start",
-                source="ClaimsAgent",
-                data={
-                    "message_id": message_id,
-                    "connection_id": connection_id,
-                },
-                traceparent=child_traceparent,
-                tracestate=child_tracestate,
-            )
-        )
-        logger.info(f"Stream started for message {message_id}")
-
-        # Call the model with streaming
         logger.info("Calling claims model with streaming")
         chunks = claims_optimized_dspy(chat_history=conversation_string, config=config)
-        chunk_count = 0
 
-        # Process the streaming chunks
-        try:
-            async for chunk in chunks:
-                if isinstance(chunk, StreamResponse):
-                    chunk_count += 1
-                    await human_stream_channel.publish(
-                        TracedMessage(
-                            type="agent_message_stream_chunk",
-                            source="ClaimsAgent",
-                            data={
-                                "message_chunk": chunk.chunk,
-                                "message_id": message_id,
-                                "chunk_index": chunk_count,
-                                "connection_id": connection_id,
-                            },
-                            traceparent=child_traceparent,
-                            tracestate=child_tracestate,
-                        )
-                    )
-                elif isinstance(chunk, Prediction):
-                    # Get the complete response
-                    response = chunk.final_response
-                    if response:
-                        response = response.replace(" [[ ## completed ## ]]", "")
-
-                    logger.info(
-                        f"Sending stream end with response: {response[:100] if response else 'EMPTY'}"
-                    )
-                    await human_stream_channel.publish(
-                        TracedMessage(
-                            type="agent_message_stream_end",
-                            source="ClaimsAgent",
-                            data={
-                                "message_id": message_id,
-                                "message": response,
-                                "agent": "ClaimsAgent",
-                                "connection_id": connection_id,
-                            },
-                            traceparent=child_traceparent,
-                            tracestate=child_tracestate,
-                        )
-                    )
-                    logger.info(f"Stream ended for message {message_id}")
-        except Exception as e:
-            logger.error(f"Error in streaming response: {e}", exc_info=True)
-            # Send an error message to end the stream
-            await human_stream_channel.publish(
-                TracedMessage(
-                    type="agent_message_stream_end",
-                    source="ClaimsAgent",
-                    data={
-                        "message_id": message_id,
-                        "message": "I'm sorry, I encountered an error while processing your request.",
-                        "agent": "ClaimsAgent",
-                        "connection_id": connection_id,
-                    },
-                    traceparent=child_traceparent,
-                    tracestate=child_tracestate,
-                )
-            )
+        await process_request_stream(
+            chunks,
+            agent_name="ClaimsAgent",
+            connection_id=connection_id,
+            message_id=message_id,
+            human_stream_channel=human_stream_channel,
+            tracer=tracer,
+            logger=logger,
+            child_traceparent=child_traceparent,
+            child_tracestate=child_tracestate,
+        )
 
 
 @claims_agent.subscribe(
@@ -167,13 +96,22 @@ async def process_claims_request(
 )
 @traced_handler("handle_claim_request")
 async def handle_claim_request(msg: TracedMessage) -> None:
-    """Handle incoming claim request messages from the agents channel."""
+    """Process a claim request from the agents channel.
+
+    Args:
+        msg: The traced message containing chat history and metadata.
+
+    Returns:
+        None
+    """
     try:
         chat_messages: List[ChatMessage] = msg.data.get("chat_messages", [])
         connection_id: str = msg.data.get("connection_id", "unknown")
 
         if not chat_messages:
             logger.warning(f"Empty chat history for connection: {connection_id}")
+            carrier = {}
+            TraceContextTextMapPropagator().inject(carrier)
             await human_channel.publish(
                 TracedMessage(
                     type="agent_message",
@@ -183,63 +121,87 @@ async def handle_claim_request(msg: TracedMessage) -> None:
                         "connection_id": connection_id,
                         "agent": "ClaimsAgent",
                     },
-                    traceparent=msg.traceparent,
-                    tracestate=msg.tracestate,
+                    traceparent=carrier.get("traceparent"),
+                    tracestate=carrier.get("tracestate", ""),
                 )
             )
             return
 
-        conversation_string = get_conversation_string(chat_messages)
+        conversation_string = format_conversation(
+            chat_messages, tracer=tracer, logger=logger
+        )
         logger.info(f"Processing claim request for connection {connection_id}")
 
         await process_claims_request(
-            conversation_string, connection_id, str(msg.id), timeout_seconds=30.0
+            conversation_string,
+            connection_id,
+            str(msg.id),
+            timeout_seconds=settings.request_timeout_seconds,
         )
 
-    except Exception as e:
-        logger.error(f"Error in ClaimsAgent: {e}", exc_info=True)
+    except ValueError as exc:
+        logger.warning("Invalid claims request: %s", exc, exc_info=True)
+        carrier = {}
+        TraceContextTextMapPropagator().inject(carrier)
+        await human_channel.publish(
+            TracedMessage(
+                type="agent_message",
+                source="ClaimsAgent",
+                data={
+                    "message": str(exc),
+                    "connection_id": connection_id,
+                    "agent": "ClaimsAgent",
+                },
+                traceparent=carrier.get("traceparent"),
+                tracestate=carrier.get("tracestate", ""),
+            )
+        )
+    except (ConnectionError, asyncio.TimeoutError) as exc:
+        logger.error("Network error in ClaimsAgent: %s", exc, exc_info=True)
+        carrier = {}
+        TraceContextTextMapPropagator().inject(carrier)
+        await human_channel.publish(
+            TracedMessage(
+                type="agent_message",
+                source="ClaimsAgent",
+                data={
+                    "message": "I'm having network issues processing your request. Please try again.",
+                    "connection_id": connection_id,
+                    "agent": "ClaimsAgent",
+                },
+                traceparent=carrier.get("traceparent"),
+                tracestate=carrier.get("tracestate", ""),
+            )
+        )
+    except Exception as exc:
+        logger.error(f"Unhandled error in ClaimsAgent: {exc}", exc_info=True)
+        carrier = {}
+        TraceContextTextMapPropagator().inject(carrier)
         await human_channel.publish(
             TracedMessage(
                 type="agent_message",
                 source="ClaimsAgent",
                 data={
                     "message": "I apologize, but I'm having trouble processing your request right now. Please try again.",
-                    "connection_id": locals().get("connection_id", "unknown"),
+                    "connection_id": connection_id,
                     "agent": "ClaimsAgent",
                 },
-                traceparent=msg.traceparent if "msg" in locals() else None,
-                tracestate=msg.tracestate if "msg" in locals() else None,
+                traceparent=carrier.get("traceparent"),
+                tracestate=carrier.get("tracestate", ""),
             )
         )
 
 
 @claims_agent.subscribe(channel=agents_channel)
 async def handle_other_messages(msg: TracedMessage) -> None:
-    """Handle non-claim messages received on the agent channel."""
+    """Handle non-claim messages received on the agent channel.
+
+    Args:
+        msg: Incoming message that does not match this agent's filter.
+
+    Returns:
+        None
+    """
     logger.debug("Received non-claim message: %s", msg)
 
 
-if __name__ == "__main__":
-
-    async def run():
-        from libraries.dspy_set_language_model import dspy_set_language_model
-
-        dspy_set_language_model(settings)
-
-        await clear_channels()
-
-        test_conversation = (
-            "User: Hi, I'd like to file a new claim.\n"
-            "ClaimsAgent: Certainly! Could you provide your policy number and incident details?\n"
-            "User: Policy A12345, my car was hit at a stop sign.\n"
-        )
-
-        logger.info("Running test query for claims agent")
-        chunks = claims_optimized_dspy(chat_history=test_conversation)
-        async for chunk in chunks:
-            if isinstance(chunk, StreamResponse):
-                logger.info(f"Chunk: {chunk.chunk}")
-            elif isinstance(chunk, Prediction):
-                logger.info(f"Final response: {chunk.final_response}")
-
-    asyncio.run(run())
