+++ examples/multi_agent_human_chat/agents/escalation/main.py
@@ -62,5 +62,7 @@ if __name__ == "__main__":
         asyncio.run(main())
     except KeyboardInterrupt:
         logger.info("Shutting down escalation agent")
+    except OSError as e:
+        logger.error(f"Server error: {e}", exc_info=True)
     except Exception as e:
         logger.error(f"Error in main: {e}", exc_info=True)
