+++ examples/multi_agent_human_chat/agents/frontend/__init__.py
@@ -1,0 +1,5 @@
+"""Frontend agent package."""
+
+from .agent import add_websocket_gateway, frontend_agent
+
+__all__ = ["add_websocket_gateway", "frontend_agent"]
