+++ examples/multi_agent_human_chat/agents/policies/agent/agent.py
@@ -1,19 +1,18 @@
 import asyncio
 from typing import List
 
-from dspy import Prediction
-from dspy.streaming import StreamResponse
 from eggai import Agent, Channel
+from opentelemetry.propagate import TraceContextTextMapPropagator
 
 from agents.policies.agent.react import policies_react_dspy
 from agents.policies.config import settings
 from agents.policies.types import ChatMessage, ModelConfig
-from libraries.channels import channels, clear_channels
+from libraries.channels import channels
 from libraries.logger import get_console_logger
+from libraries.streaming_agent import format_conversation, process_request_stream
 from libraries.tracing import (
     TracedMessage,
-    create_tracer,
-    format_span_as_traceparent,
+    get_tracer,
     traced_handler,
 )
 from libraries.tracing.init_metrics import init_token_metrics
@@ -24,38 +23,11 @@ logger = get_console_logger("policies_agent.handler")
 agents_channel = Channel(channels.agents)
 human_channel = Channel(channels.human)
 human_stream_channel = Channel(channels.human_stream)
-tracer = create_tracer("policies_agent")
+tracer = get_tracer("policies_agent")
 
 init_token_metrics(
     port=settings.prometheus_metrics_port, application_name=settings.app_name
 )
-
-
-def get_conversation_string(chat_messages: List[ChatMessage]) -> str:
-    """Format chat messages into a conversation string."""
-    with tracer.start_as_current_span("get_conversation_string") as span:
-        safe_set_attribute(
-            span, "chat_messages_count", len(chat_messages) if chat_messages else 0
-        )
-
-        if not chat_messages:
-            safe_set_attribute(span, "empty_messages", True)
-            return ""
-
-        conversation_parts = []
-        for chat in chat_messages:
-            if "content" not in chat:
-                safe_set_attribute(span, "invalid_message", True)
-                logger.warning("Message missing content field")
-                continue
-
-            role = chat.get("role", "User")
-            conversation_parts.append(f"{role}: {chat['content']}")
-
-        conversation = "\n".join(conversation_parts) + "\n"
-        safe_set_attribute(span, "conversation_length", len(conversation))
-        return conversation
-
 
 async def process_policy_request(
     conversation_string: str,
@@ -67,97 +39,37 @@ async def process_policy_request(
     # Create model config with timeout value
     config = ModelConfig(timeout_seconds=timeout_seconds or 30.0)
     with tracer.start_as_current_span("process_policy_request") as span:
-        child_traceparent, child_tracestate = format_span_as_traceparent(span)
+        carrier = {}
+        TraceContextTextMapPropagator().inject(carrier)
+        child_traceparent = carrier.get("traceparent")
+        child_tracestate = carrier.get("tracestate", "")
         safe_set_attribute(span, "connection_id", connection_id)
         safe_set_attribute(span, "message_id", message_id)
         safe_set_attribute(span, "conversation_length", len(conversation_string))
         safe_set_attribute(span, "timeout_seconds", config.timeout_seconds)
 
-        if not conversation_string or len(conversation_string.strip()) < 5:
+        if (
+            not conversation_string
+            or len(conversation_string.strip()) < settings.min_conversation_length
+        ):
             safe_set_attribute(span, "error", "Empty or too short conversation")
             span.set_status(1, "Invalid input")
             raise ValueError("Conversation history is too short to process")
 
-        # Start the stream
-        await human_stream_channel.publish(
-            TracedMessage(
-                type="agent_message_stream_start",
-                source="PoliciesAgent",
-                data={
-                    "message_id": message_id,
-                    "connection_id": connection_id,
-                },
-                traceparent=child_traceparent,
-                tracestate=child_tracestate,
-            )
-        )
-        logger.info(f"Stream started for message {message_id}")
-
-        # Call the model with streaming
         logger.info("Calling policies model with streaming")
         chunks = policies_react_dspy(chat_history=conversation_string, config=config)
-        chunk_count = 0
 
-        # Process the streaming chunks
-        try:
-            async for chunk in chunks:
-                if isinstance(chunk, StreamResponse):
-                    chunk_count += 1
-                    await human_stream_channel.publish(
-                        TracedMessage(
-                            type="agent_message_stream_chunk",
-                            source="PoliciesAgent",
-                            data={
-                                "message_chunk": chunk.chunk,
-                                "message_id": message_id,
-                                "chunk_index": chunk_count,
-                                "connection_id": connection_id,
-                            },
-                            traceparent=child_traceparent,
-                            tracestate=child_tracestate,
-                        )
-                    )
-                elif isinstance(chunk, Prediction):
-                    # Get the complete response
-                    response = chunk.final_response
-                    if response:
-                        response = response.replace(" [[ ## completed ## ]]", "")
-
-                    logger.info(
-                        f"Sending stream end with response: {response[:100] if response else 'EMPTY'}"
-                    )
-                    await human_stream_channel.publish(
-                        TracedMessage(
-                            type="agent_message_stream_end",
-                            source="PoliciesAgent",
-                            data={
-                                "message_id": message_id,
-                                "message": response,
-                                "agent": "PoliciesAgent",
-                                "connection_id": connection_id,
-                            },
-                            traceparent=child_traceparent,
-                            tracestate=child_tracestate,
-                        )
-                    )
-                    logger.info(f"Stream ended for message {message_id}")
-        except Exception as e:
-            logger.error(f"Error in streaming response: {e}", exc_info=True)
-            # Send an error message to end the stream
-            await human_stream_channel.publish(
-                TracedMessage(
-                    type="agent_message_stream_end",
-                    source="PoliciesAgent",
-                    data={
-                        "message_id": message_id,
-                        "message": "I'm sorry, I encountered an error while processing your request.",
-                        "agent": "PoliciesAgent",
-                        "connection_id": connection_id,
-                    },
-                    traceparent=child_traceparent,
-                    tracestate=child_tracestate,
-                )
-            )
+        await process_request_stream(
+            chunks,
+            agent_name="PoliciesAgent",
+            connection_id=connection_id,
+            message_id=message_id,
+            human_stream_channel=human_stream_channel,
+            tracer=tracer,
+            logger=logger,
+            child_traceparent=child_traceparent,
+            child_tracestate=child_tracestate,
+        )
 
 
 @policies_agent.subscribe(
@@ -169,12 +81,14 @@ async def process_policy_request(
 @traced_handler("handle_policy_request")
 async def handle_policy_request(msg: TracedMessage) -> None:
     """Handle incoming policy request messages from the agents channel."""
+    connection_id = msg.data.get("connection_id", "unknown")
     try:
         chat_messages: List[ChatMessage] = msg.data.get("chat_messages", [])
-        connection_id: str = msg.data.get("connection_id", "unknown")
 
         if not chat_messages:
             logger.warning(f"Empty chat history for connection: {connection_id}")
+            carrier = {}
+            TraceContextTextMapPropagator().inject(carrier)
             await human_channel.publish(
                 TracedMessage(
                     type="agent_message",
@@ -184,32 +98,70 @@ async def handle_policy_request(msg: TracedMessage) -> None:
                         "connection_id": connection_id,
                         "agent": "PoliciesAgent",
                     },
-                    traceparent=msg.traceparent,
-                    tracestate=msg.tracestate,
+                    traceparent=carrier.get("traceparent"),
+                    tracestate=carrier.get("tracestate", ""),
                 )
             )
             return
 
-        conversation_string = get_conversation_string(chat_messages)
+        conversation_string = format_conversation(
+            chat_messages, tracer=tracer, logger=logger
+        )
         logger.info(f"Processing policy request for connection {connection_id}")
 
         await process_policy_request(
             conversation_string, connection_id, str(msg.id), timeout_seconds=30.0
         )
 
-    except Exception as e:
-        logger.error(f"Error in PoliciesAgent: {e}", exc_info=True)
+    except ValueError as exc:
+        logger.warning("Invalid policies request: %s", exc, exc_info=True)
+        carrier = {}
+        TraceContextTextMapPropagator().inject(carrier)
+        await human_channel.publish(
+            TracedMessage(
+                type="agent_message",
+                source="PoliciesAgent",
+                data={
+                    "message": str(exc),
+                    "connection_id": connection_id,
+                    "agent": "PoliciesAgent",
+                },
+                traceparent=carrier.get("traceparent"),
+                tracestate=carrier.get("tracestate", ""),
+            )
+        )
+    except (ConnectionError, asyncio.TimeoutError) as exc:
+        logger.error("Network error in PoliciesAgent: %s", exc, exc_info=True)
+        carrier = {}
+        TraceContextTextMapPropagator().inject(carrier)
+        await human_channel.publish(
+            TracedMessage(
+                type="agent_message",
+                source="PoliciesAgent",
+                data={
+                    "message": "I'm having network issues processing your request. Please try again.",
+                    "connection_id": connection_id,
+                    "agent": "PoliciesAgent",
+                },
+                traceparent=carrier.get("traceparent"),
+                tracestate=carrier.get("tracestate", ""),
+            )
+        )
+    except Exception as exc:
+        logger.error(f"Unhandled error in PoliciesAgent: {exc}", exc_info=True)
+        carrier = {}
+        TraceContextTextMapPropagator().inject(carrier)
         await human_channel.publish(
             TracedMessage(
                 type="agent_message",
                 source="PoliciesAgent",
                 data={
                     "message": "I apologize, but I'm having trouble processing your request right now. Please try again.",
-                    "connection_id": locals().get("connection_id", "unknown"),
+                    "connection_id": connection_id,
                     "agent": "PoliciesAgent",
                 },
-                traceparent=msg.traceparent if "msg" in locals() else None,
-                tracestate=msg.tracestate if "msg" in locals() else None,
+                traceparent=carrier.get("traceparent"),
+                tracestate=carrier.get("tracestate", ""),
             )
         )
 
@@ -220,27 +172,3 @@ async def handle_other_messages(msg: TracedMessage) -> None:
     logger.debug("Received non-policy message: %s", msg)
 
 
-if __name__ == "__main__":
-
-    async def run():
-        from libraries.dspy_set_language_model import dspy_set_language_model
-
-        dspy_set_language_model(settings)
-
-        await clear_channels()
-
-        test_conversation = (
-            "User: I need information about my policy.\n"
-            "PoliciesAgent: Sure, I can help with that. Could you please provide me with your policy number?\n"
-            "User: My policy number is A12345\n"
-        )
-
-        logger.info("Running test query for policies agent")
-        chunks = policies_react_dspy(chat_history=test_conversation)
-        async for chunk in chunks:
-            if isinstance(chunk, StreamResponse):
-                logger.info(f"Chunk: {chunk.chunk}")
-            elif isinstance(chunk, Prediction):
-                logger.info(f"Final response: {chunk.final_response}")
-
-    asyncio.run(run())
