+++ examples/multi_agent_human_chat/libraries/__init__.py
@@ -4,4 +4,21 @@ Common libraries for the multi-agent human chat example.
 
 from libraries.patches import patch_usage_tracker
 
+from .streaming_agent import format_conversation, process_request_stream
+from .trace_utils import (
+    get_message_content,
+    get_message_id,
+    get_message_metadata,
+    propagate_trace_context,
+)
+
 patch_usage_tracker()
+
+__all__ = [
+    "get_message_metadata",
+    "get_message_content",
+    "get_message_id",
+    "propagate_trace_context",
+    "process_request_stream",
+    "format_conversation",
+]
